refactor: clearer names for auth components
